reorganise module for future additions
