+++ ctaplot/__init__.py
@@ -1,7 +1,7 @@
-from .plots import *
-from .ana import *
-from .dataset import *
-from .io import *
+from ctaplot.plots.plots import *
+from ctaplot.ana.ana import *
+from ctaplot.io.dataset import *
+from ctaplot.io.io import *
 from .gammaboard import gammaboard
 
 __version__ = '0.4.0'
