+++ ctaplot/tests/test_imports.py
@@ -1,18 +1,18 @@
 def test_import_plots():
     try:
-        from ctaplot import plots
+        from ctaplot.plots import plots
     except:
         raise()
 
 def test_import_ana():
     try:
-        from ctaplot import ana
+        from ctaplot.ana import ana
     except:
         raise ()
 
 def test_import_dataset():
     try:
-        from ctaplot import dataset
+        from ctaplot.io import dataset
     except:
         raise ()
